move instantiator to an independent crate
