+++ pallets/dispenser/Cargo.toml
@@ -30,7 +30,6 @@ frame-support.workspace = true
 frame-system.workspace = true
 
 polimec-common.workspace = true
-polimec-common-test-utils.workspace = true
 sp-std.workspace = true
 sp-runtime.workspace = true
 
@@ -41,6 +40,7 @@ sp-runtime.workspace = true
 pallet-balances.workspace = true
 pallet-timestamp.workspace = true
 pallet-vesting.workspace = true
+polimec-instantiator.workspace = true
 
 [features]
 default = [ "std" ]
@@ -52,8 +52,6 @@ std = [
 	"pallet-timestamp/std",
 	"pallet-vesting/std",
 	"parity-scale-codec/std",
-	"polimec-common-test-utils/std",
-	"polimec-common/std",
 	"scale-info/std",
 	"sp-core/std",
 	"sp-io/std",
@@ -67,8 +65,6 @@ runtime-benchmarks = [
 	"pallet-balances/runtime-benchmarks",
 	"pallet-timestamp/runtime-benchmarks",
 	"pallet-vesting/runtime-benchmarks",
-	"polimec-common-test-utils/runtime-benchmarks",
-	"polimec-common/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 ]
 
@@ -78,7 +74,5 @@ try-runtime = [
 	"pallet-balances/try-runtime",
 	"pallet-timestamp/try-runtime",
 	"pallet-vesting/try-runtime",
-	"polimec-common-test-utils/try-runtime",
-	"polimec-common/try-runtime",
 	"sp-runtime/try-runtime",
 ]
