+++ polimec-common/Cargo.toml
@@ -27,6 +27,10 @@ sp-std.workspace = true
 sp-runtime.workspace = true
 itertools.workspace = true
 xcm.workspace = true
+reqwest = { version = "0.11.4", features = [
+	"blocking",
+	"default-tls",
+], default-features = false, optional = true }
 
 
 [features]
@@ -45,6 +49,7 @@ std = [
 	"sp-runtime/std",
 	"sp-std/std",
 	"xcm/std",
+	"reqwest",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
