--- polimec-common/test-utils/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = 'polimec-common-test-utils'
-description = 'Testing helpers for the polimec-common crate'
-authors = ['Polimec Foundation <info@polimec.org>']
-documentation = "https://wiki.polimec.org/"
-edition = "2021"
-homepage = "https://www.polimec.org/"
-license-file = "LICENSE"
-readme = "README.md"
-repository = "https://github.com/Polimec/polimec-node"
-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-parity-scale-codec = { workspace = true, features = ["derive"] }
-jwt-compact = { workspace = true, features = [
-    "ed25519-dalek",
-], default-features = false }
-polimec-common = { workspace = true, default-features = false }
-reqwest = { version = "0.11.4", features = [
-    "blocking",
-	"default-tls",
-], default-features = false, optional = true }
-frame-support.workspace = true
-chrono = { version = "0.4.19", default-features = false }
-
-[features]
-default = [ "std" ]
-std = [
-	"chrono/std",
-	"frame-support/std",
-	"jwt-compact/std",
-	"parity-scale-codec/std",
-	"polimec-common/std",
-	"reqwest",
-]
-try-runtime = [ "frame-support/try-runtime", "polimec-common/try-runtime" ]
-runtime-benchmarks = [
-	"frame-support/runtime-benchmarks",
-	"polimec-common/runtime-benchmarks",
-]
